Possible fix for "double starting of services" issue.
Only an issue for ConstructOnDemand services that are
directly referenced in the init functions of the scenario
script. This fix just checks that the service doesn't
already exist before it starts it.

Other possible fixes are to change the initialization order
so that these services are initialized before the scenario
script is loaded (don't know if this would effect existing
services). Could also check here for construct on demand
services and just skip initializing them in this loop.
This might make the most sense since otherwise it's not really
construct on demand.

Lots of options for a fix here.