Fixed rendering of entities added and removed in the same tick
The EntityShapeManager caches entities that have been added or removed between
ticks. When a polygon entity is both added and removed in a single tick, the
shape manager tried to process both events and threw an exception when trying to
retrieve the waypoint components of the polygon from the simulation, because
there was no attached simulation to query.

The shape manager now skips adding any entities that are not attached to the
simulation, since we can infer that those have since been removed.