Fix entity drifting bug when using clampPointToGround
clampPointToGround would return a different vector than the one passed in, even when the vector shouldn't change.
Now if any of the x, y, or z values of the new vector are within a threshold of the passed in vector the values won't change.
This fixes a bug where entities would change position even when they had 0 velocity